fix: rename package name correctly
