+++ app/src/androidTest/java/com/github/onlynotesswent/ExampleInstrumentedTest.kt
@@ -1,8 +1,8 @@
-package com.github.onlynoteswent
+package com.github.onlynotesswent
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.onlynoteswent.screen.SecondScreen
+import com.github.onlynotesswent.screen.MainScreen
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import org.junit.Rule
@@ -15,17 +15,17 @@ import org.junit.runner.RunWith
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class SecondActivityTest : TestCase() {
+class MainActivityTest : TestCase() {
 
-  @get:Rule val composeTestRule = createAndroidComposeRule<SecondActivity>()
+  @get:Rule val composeTestRule = createAndroidComposeRule<MainActivity>()
 
   @Test
   fun test() = run {
-    step("Start Second Activity") {
-      ComposeScreen.onComposeScreen<SecondScreen>(composeTestRule) {
+    step("Start Main Activity") {
+      ComposeScreen.onComposeScreen<MainScreen>(composeTestRule) {
         simpleText {
           assertIsDisplayed()
-          assertTextEquals("Hello Robolectric!")
+          assertTextEquals("Hello Android!")
         }
       }
     }
