+++ app/src/main/java/com/github/onlynotesswent/SimpleData.kt
@@ -1,4 +1,4 @@
-package com.github.onlynoteswent
+package com.github.onlynotesswent
 
 import kotlin.math.sqrt
