+++ app/src/main/java/com/github/onlynotesswent/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.github.onlynoteswent.ui.theme
+package com.github.onlynotesswent.ui.theme
 
 import android.app.Activity
 import android.os.Build
