+++ app/src/main/java/com/github/onlynotesswent/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.onlynoteswent.ui.theme
+package com.github.onlynotesswent.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
