+++ app/src/test/java/com/github/onlynotesswent/PointTest.kt
@@ -1,4 +1,4 @@
-package com.github.onlynoteswent
+package com.github.onlynotesswent
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
